# packages/generated/api-client-csharp/api/openapi.yaml
openapi: 3.1.0
info:
  title: Plerion
  version: 0.1.0
servers:
- url: /
security:
- oauth2:
  - openid
- bearerAuth: []
paths:
...
  /reconstructions/{id}/localization_map:
    get:
      operationId: get_reconstruction_localization_map
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                format: uuid
                title: Response Get Reconstruction Localization Map
                type: string
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Reconstruction Localization Map
      tags:
      - Default
  /reconstructions/{id}/status:
    get:
      operationId: get_reconstruction_status
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                enum:
                - queued
                - pending
                - downloading
                - extracting_features
                - matching_features
                - reconstructing
                - training_opq_matrix
                - training_product_quantizer
                - uploading
                - succeeded
                - failed
                title: Response Get Reconstruction Status
                type: string
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Reconstruction Status
      tags:
      - Default
  /reconstructions/{id}/points:
    get:
      operationId: get_reconstruction_points
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/PointCloudPoint"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Reconstruction Points
      tags:
      - Default
  /reconstructions/{id}/points.ply:
    get:
      operationId: get_reconstruction_points3D_ply
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/octet-stream:
              schema:
                format: binary
                type: string
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Reconstruction Points3D Ply
      tags:
      - Default
  /reconstructions/{id}/image_poses:
    get:
      operationId: get_reconstruction_image_poses
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Transform"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Reconstruction Image Poses
      tags:
      - Default
  /localization_maps:
    delete:
      operationId: delete_localization_maps
      parameters:
      - description: List of Ids to delete
        explode: true
        in: query
        name: ids
        required: true
        schema:
          description: List of Ids to delete
          items:
            format: uuid
            type: string
            x-schema-name: delete_localization_maps_param_ids_item
            nullable: false
          type: array
        style: form
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Delete Localization Maps
      tags:
      - Default
    get:
      operationId: get_localization_maps
      parameters:
      - description: Optional list of Ids to filter by
        explode: true
        in: query
        name: ids
        required: false
        schema:
          items:
            format: uuid
            type: string
            x-schema-name: get_localization_maps_param_ids_item
            nullable: false
          type: array
          nullable: true
        style: form
      - description: Optional list of Reconstruction Ids to filter by
        explode: true
        in: query
        name: reconstruction_ids
        required: false
        schema:
          items:
            format: uuid
            type: string
            x-schema-name: get_localization_maps_param_reconstruction_ids_item
            nullable: false
          type: array
          nullable: true
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/LocalizationMapRead"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Localization Maps
      tags:
      - Default
    patch:
      operationId: update_localization_maps
      parameters:
      - explode: true
        in: query
        name: allow_missing
        required: false
        schema:
          default: false
          title: Allow Missing
          type: boolean
        style: form
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: "#/components/schemas/LocalizationMapBatchUpdate"
              type: array
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/LocalizationMapRead"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Update Localization Maps
      tags:
      - Default
    post:
      operationId: create_localization_map
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocalizationMapCreate"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocalizationMapRead"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Create Localization Map
      tags:
      - Default
  /localization_maps/{id}:
    delete:
      operationId: delete_localization_map
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Delete Localization Map
      tags:
      - Default
    get:
      operationId: get_localization_map
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocalizationMapRead"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Localization Map
      tags:
      - Default
    patch:
      operationId: update_localization_map
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Id
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocalizationMapUpdate"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocalizationMapRead"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Update Localization Map
      tags:
      - Default
  /localization_maps/{id}/points:
    get:
      operationId: get_localization_map_points
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/PointCloudPoint"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Localization Map Points
      tags:
      - Default
  /localization_maps/{id}/points.ply:
    get:
      operationId: get_localization_map_points_ply
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/octet-stream:
              schema:
                format: binary
                type: string
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Localization Map Points Ply
      tags:
      - Default
  /localization_maps/{id}/image_poses:
    patch:
      operationId: update_localization_map_image_poses
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Transform"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Update Localization Map Image Poses
      tags:
      - Default
...
  /localization_sessions/{localization_session_id}/maps:
    post:
      operationId: load_localization_maps
      parameters:
      - explode: false
        in: path
        name: localization_session_id
        required: true
        schema:
          format: uuid
          title: Localization Session Id
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              description: IDs of localization maps to load
              items:
                format: uuid
                type: string
                x-schema-name: load_localization_maps_body_application_json_item
                nullable: false
              type: array
        required: true
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Load Localization Maps
      tags:
      - Default
  /localization_sessions/{localization_session_id}/maps/{map_id}:
    delete:
      operationId: unload_map
      parameters:
      - explode: false
        in: path
        name: localization_session_id
        required: true
        schema:
          format: uuid
          title: Localization Session Id
          type: string
        style: simple
      - explode: false
        in: path
        name: map_id
        required: true
        schema:
          format: uuid
          title: Map Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Unload Map
      tags:
      - Default
...
  /localization_sessions/{localization_session_id}/localization:
    post:
      operationId: localize_image
      parameters:
      - explode: false
        in: path
        name: localization_session_id
        required: true
        schema:
          format: uuid
          title: Localization Session Id
          type: string
        style: simple
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/Body_localize_image"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/MapLocalization"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Localize Image
      tags:
      - Default
...
components:
  schemas:
    Body_localize_image:
      properties:
        image:
          format: binary
          title: Image
          type: string
      required:
      - image
      title: Body_localize_image
...
    Color:
      example:
        r: 0
        b: 1
        g: 6
      properties:
        r:
          title: R
          type: integer
        g:
          title: G
          type: integer
        b:
          title: B
          type: integer
      required:
      - b
      - g
      - r
      title: Color
...
    HTTPValidationError:
      example:
        detail:
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          type: type
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          type: type
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
      title: HTTPValidationError
...
    LocalizationMapBatchUpdate:
      example:
        rotation_x: 5.637376656633329
        color: 9
        rotation_w: 7.061401241503109
        active: true
        rotation_z: 2.3021358869347655
        rotation_y: 0.8008281904610115
        lighting: 3
        position_x: 6.027456183070403
        position_y: 1.4658129805029452
        position_z: 5.962133916683182
        name: name
        reconstruction_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        reconstruction_id:
          format: uuid
          nullable: true
          type: string
        id:
          format: uuid
          title: Id
          type: string
        rotation_y:
          nullable: true
          type: number
        position_x:
          nullable: true
          type: number
        position_y:
          nullable: true
          type: number
        position_z:
          nullable: true
          type: number
        rotation_x:
          nullable: true
          type: number
        rotation_z:
          nullable: true
          type: number
        rotation_w:
          nullable: true
          type: number
        color:
          nullable: true
          type: integer
        active:
          nullable: true
          type: boolean
        lighting:
          nullable: true
          type: integer
        name:
          nullable: true
          type: string
      required:
      - id
      title: LocalizationMapBatchUpdate
    LocalizationMapCreate:
      example:
        rotation_x: 5.637376656633329
        color: 9
        rotation_w: 7.061401241503109
        active: true
        rotation_z: 2.3021358869347655
        rotation_y: 0.8008281904610115
        lighting: 3
        position_x: 6.027456183070403
        position_y: 1.4658129805029452
        position_z: 5.962133916683182
        name: name
        reconstruction_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        reconstruction_id:
          format: uuid
          title: Reconstruction Id
          type: string
        id:
          format: uuid
          nullable: true
          type: string
        rotation_y:
          title: Rotation Y
          type: number
        position_x:
          title: Position X
          type: number
        position_y:
          title: Position Y
          type: number
        position_z:
          title: Position Z
          type: number
        rotation_x:
          title: Rotation X
          type: number
        rotation_z:
          title: Rotation Z
          type: number
        rotation_w:
          title: Rotation W
          type: number
        color:
          title: Color
          type: integer
        active:
          nullable: true
          type: boolean
        lighting:
          nullable: true
          type: integer
        name:
          nullable: true
          type: string
      required:
      - color
      - position_x
      - position_y
      - position_z
      - reconstruction_id
      - rotation_w
      - rotation_x
      - rotation_y
      - rotation_z
      title: LocalizationMapCreate
    LocalizationMapRead:
      example:
        rotation_x: 5.637376656633329
        color: 9
        rotation_w: 7.061401241503109
        created_at: 2000-01-23T04:56:07.000+00:00
        active: true
        rotation_z: 2.3021358869347655
        rotation_y: 0.8008281904610115
        lighting: 3
        position_x: 6.027456183070403
        position_y: 1.4658129805029452
        position_z: 5.962133916683182
        updated_at: 2000-01-23T04:56:07.000+00:00
        name: name
        reconstruction_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        reconstruction_id:
          format: uuid
          title: Reconstruction Id
          type: string
        id:
          format: uuid
          title: Id
          type: string
        rotation_y:
          title: Rotation Y
          type: number
        position_x:
          title: Position X
          type: number
        position_y:
          title: Position Y
          type: number
        position_z:
          title: Position Z
          type: number
        rotation_x:
          title: Rotation X
          type: number
        updated_at:
          format: date-time
          title: Updated At
          type: string
        rotation_z:
          title: Rotation Z
          type: number
        rotation_w:
          title: Rotation W
          type: number
        color:
          title: Color
          type: integer
        created_at:
          format: date-time
          title: Created At
          type: string
        active:
          title: Active
          type: boolean
        lighting:
          nullable: true
          type: integer
        name:
          nullable: true
          type: string
      required:
      - active
      - color
      - created_at
      - id
      - position_x
      - position_y
      - position_z
      - reconstruction_id
      - rotation_w
      - rotation_x
      - rotation_y
      - rotation_z
      - updated_at
      title: LocalizationMapRead
    LocalizationMapUpdate:
      example:
        position_x: 6.027456183070403
        position_y: 1.4658129805029452
        position_z: 5.962133916683182
        rotation_x: 5.637376656633329
        color: 9
        rotation_w: 7.061401241503109
        name: name
        reconstruction_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        active: true
        rotation_z: 2.3021358869347655
        rotation_y: 0.8008281904610115
        lighting: 3
      properties:
        reconstruction_id:
          format: uuid
          nullable: true
          type: string
        rotation_y:
          nullable: true
          type: number
        position_x:
          nullable: true
          type: number
        position_y:
          nullable: true
          type: number
        position_z:
          nullable: true
          type: number
        rotation_x:
          nullable: true
          type: number
        rotation_z:
          nullable: true
          type: number
        rotation_w:
          nullable: true
          type: number
        color:
          nullable: true
          type: integer
        active:
          nullable: true
          type: boolean
        lighting:
          nullable: true
          type: integer
        name:
          nullable: true
          type: string
      title: LocalizationMapUpdate
    LocalizationMetrics:
      example:
        image_coverage: 1.4658129805029452
        depth_z_10th_percentile: 5.637376656633329
        inlier_ratio: 0.8008281904610115
        reprojection_error_median: 6.027456183070403
        depth_z_90th_percentile: 5.962133916683182
      properties:
        inlier_ratio:
          title: Inlier Ratio
          type: number
        reprojection_error_median:
          title: Reprojection Error Median
          type: number
        image_coverage:
          title: Image Coverage
          type: number
        depth_z_90th_percentile:
          title: Depth Z 90Th Percentile
          type: number
        depth_z_10th_percentile:
          title: Depth Z 10Th Percentile
          type: number
      required:
      - depth_z_10th_percentile
      - depth_z_90th_percentile
      - image_coverage
      - inlier_ratio
      - reprojection_error_median
      title: LocalizationMetrics
...
    MapLocalization:
      example:
        transform:
          rotation:
            w: 5.962133916683182
            x: 0.8008281904610115
            "y": 6.027456183070403
            z: 1.4658129805029452
          position:
            x: 5.637376656633329
            "y": 2.3021358869347655
            z: 7.061401241503109
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        metrics:
          image_coverage: 1.4658129805029452
          depth_z_10th_percentile: 5.637376656633329
          inlier_ratio: 0.8008281904610115
          reprojection_error_median: 6.027456183070403
          depth_z_90th_percentile: 5.962133916683182
        map_transform:
          rotation:
            w: 5.962133916683182
            x: 0.8008281904610115
            "y": 6.027456183070403
            z: 1.4658129805029452
          position:
            x: 5.637376656633329
            "y": 2.3021358869347655
            z: 7.061401241503109
      properties:
        id:
          format: uuid
          title: Id
          type: string
        transform:
          $ref: "#/components/schemas/Transform"
        map_transform:
          $ref: "#/components/schemas/Transform"
        metrics:
          $ref: "#/components/schemas/LocalizationMetrics"
      required:
      - id
      - map_transform
      - metrics
      - transform
      title: MapLocalization
...
    PointCloudPoint:
      example:
        color:
          r: 0
          b: 1
          g: 6
        position:
          x: 5.637376656633329
          "y": 2.3021358869347655
          z: 7.061401241503109
      properties:
        position:
          $ref: "#/components/schemas/Vector3"
        color:
          $ref: "#/components/schemas/Color"
      required:
      - color
      - position
      title: PointCloudPoint
    Quaternion:
      example:
        w: 5.962133916683182
        x: 0.8008281904610115
        "y": 6.027456183070403
        z: 1.4658129805029452
      properties:
        x:
          title: X
          type: number
        "y":
          title: "Y"
          type: number
        z:
          title: Z
          type: number
        w:
          title: W
          type: number
      required:
      - w
      - x
      - "y"
      - z
      title: Quaternion
...
    Transform:
      example:
        rotation:
          w: 5.962133916683182
          x: 0.8008281904610115
          "y": 6.027456183070403
          z: 1.4658129805029452
        position:
          x: 5.637376656633329
          "y": 2.3021358869347655
          z: 7.061401241503109
      properties:
        position:
          $ref: "#/components/schemas/Vector3"
        rotation:
          $ref: "#/components/schemas/Quaternion"
      required:
      - position
      - rotation
      title: Transform
    ValidationError:
      example:
        msg: msg
        loc:
        - ValidationError_loc_inner
        - ValidationError_loc_inner
        type: type
      properties:
        loc:
          items:
            $ref: "#/components/schemas/ValidationError_loc_inner"
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
    Vector3:
      example:
        x: 5.637376656633329
        "y": 2.3021358869347655
        z: 7.061401241503109
      properties:
        x:
          title: X
          type: number
        "y":
          title: "Y"
          type: number
        z:
          title: Z
          type: number
      required:
      - x
      - "y"
      - z
      title: Vector3
...
    ValidationError_loc_inner:
      anyOf:
      - type: string
      - type: integer
  securitySchemes: {}